requirements.txt
pandas>=2.0.0
pyarrow>=14.0.0
requests>=2.31.0
python-dotenv>=1.0.0
web3>=6.11.0
eth-account>=0.10.0
google-genai>=0.3.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
pydantic>=2.5.0
joblib>=1.3.0
numpy>=1.24.0
